+++ packages/video_player/video_player_web/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## NEXT
 
+* Updates the minimum supported `package:web` version to 1.0.0.
 * Updates minimum supported SDK version to Flutter 3.38/Dart 3.10.
 
 ## 2.4.0